ファイル構成変更

+++ CajonFW/SW.h
@@ -1,7 +1,7 @@
 #ifndef _SW_H_
 #define _SW_H_
 
-#include <FreeRTOS.h>
+#include <M5Unified.h>
 #include <queue.h>
 
 // キューの定義
@@ -13,9 +13,9 @@
 
 // 要求の構造体
 typedef struct StagSWParam {
-  int unShortPushReq;
+  uint16_t unShortPushReq;
   QueueHandle_t pstShorQue;
-  int unLongPushReq;
+  uint16_t unLongPushReq;
   QueueHandle_t pstLongQue;
   TickType_t xTimeNow; 
 }TS_SWParam;
